+++ javascript/src/jobProcessor.ts
@@ -3,7 +3,6 @@ import {
   ConnectionOptions,
   AckPolicy,
   connect,
-  ConsumerInfo,
   DeliverPolicy,
   DiscardPolicy,
   JsMsg,
@@ -11,28 +10,39 @@ import {
   ReplayPolicy,
   RetentionPolicy,
   StorageType,
+  Nanos,
 } from 'nats'
 import { nanos, defer, getNextBackoff, nanosToMs } from './util'
 import _debug from 'debug'
 import { Deferred, JobDef } from './types'
+import _ from 'lodash/fp'
 
 const debug = _debug('nats-jobs')
 
+const streamDefaults = (def: JobDef) =>
+  _.defaults(
+    {
+      name: def.stream,
+      retention: RetentionPolicy.Workqueue,
+      storage: StorageType.File,
+      max_age: nanos('1w'),
+      num_replicas: 1,
+      subjects: [def.stream],
+      discard: DiscardPolicy.Old,
+      deny_delete: false,
+      deny_purge: false,
+    },
+    def.streamConfig
+  )
+
+/**
+ * Create the stream. By default we create a work queue with
+ * file storage. The default subject is the name of the stream.
+ */
 const createStream = async (conn: NatsConnection, def: JobDef) => {
   const jsm = await conn.jetstreamManager()
   // Stream config
-  const config = {
-    name: def.stream,
-    retention: RetentionPolicy.Workqueue,
-    storage: StorageType.File,
-    max_age: nanos('1w'),
-    num_replicas: 1,
-    subjects: [def.stream],
-    discard: DiscardPolicy.Old,
-    deny_delete: false,
-    deny_purge: false,
-    ...def.streamConfig,
-  }
+  const config = streamDefaults(def)
   debug('stream config %O', config)
   // Add stream
   return jsm.streams.add(config)
@@ -40,17 +50,26 @@ const createStream = async (conn: NatsConnection, def: JobDef) => {
 
 const defaultAckWait = nanos('10s')
 
+const consumerDefaults = (def: JobDef) =>
+  _.defaults(
+    {
+      durable_name: `${def.stream}Consumer`,
+      max_deliver: def.numAttempts ?? 5,
+      ack_policy: AckPolicy.Explicit,
+      ack_wait: defaultAckWait,
+      deliver_policy: DeliverPolicy.All,
+      replay_policy: ReplayPolicy.Instant,
+    },
+    def.consumerConfig
+  )
+
+/**
+ * Create a pull consumer on the stream. Require manual acks.
+ * By default we don't filter subjects.
+ */
 const createConsumer = (conn: NatsConnection, def: JobDef) => {
   // Consumer config
-  const config = {
-    durable_name: `${def.stream}Consumer`,
-    max_deliver: def.numAttempts ?? 5,
-    ack_policy: AckPolicy.Explicit,
-    ack_wait: defaultAckWait,
-    deliver_policy: DeliverPolicy.All,
-    replay_policy: ReplayPolicy.Instant,
-    ...def.consumerConfig,
-  }
+  const config = consumerDefaults(def)
   debug('consumer config %O', config)
   const js = conn.jetstream()
   // Create a pull consumer
@@ -66,7 +85,7 @@ const extendAckTimeoutThresholdFactor = 0.75
  * Automatically extend the ack timeout by periodically telling NATS
  * we're working.
  */
-const extendAckTimeout = (ackWait: number, msg: JsMsg): NodeJS.Timer => {
+const extendAckTimeout = (ackWait: Nanos, msg: JsMsg): NodeJS.Timer => {
   const intervalMs = nanosToMs(ackWait) * extendAckTimeoutThresholdFactor
   return setInterval(() => {
     debug('extend ack - wait: %d msg: %O', ackWait, msg.info)
@@ -75,25 +94,28 @@ const extendAckTimeout = (ackWait: number, msg: JsMsg): NodeJS.Timer => {
 }
 
 /**
- * Call perform for each message received on the stream.
+ * Start processing jobs based on def. To gracefully shutdown
+ * see stop method.
  */
 export const jobProcessor = async (opts?: ConnectionOptions) => {
+  // Connect to NATS
   const conn = await connect(opts)
   const js = conn.jetstream()
+  const abortController = new AbortController()
   let timer: NodeJS.Timer
   let deferred: Deferred<void>
-  const abortController = new AbortController()
 
   /**
-   * Start processing jobs based on def.
-   * To gracefully shutdown see stop method.
+   * Call perform for each message received on the stream.
    */
   const start = async (def: JobDef) => {
-    let extendAckTimer: NodeJS.Timer | undefined
     debug('job def %O', def)
     const pullInterval = def.pullInterval ?? ms('1s')
+    // Retry a failed message after a second by default
     const backoff = def.backoff ?? ms('1s')
+    // Pull down 10 messages by default
     const batch = def.batch ?? 10
+    // Automatically extend the ack timeout by default
     const autoExtendAckTimeout = def.autoExtendAckTimeout ?? true
     // Create stream
     // TODO: Maybe handle errors better
@@ -101,8 +123,7 @@ export const jobProcessor = async (opts?: ConnectionOptions) => {
     // Create pull consumer
     const ps = await createConsumer(conn, def)
     // Get consumer info
-    const consumerInfo = await ps.consumerInfo()
-    const ackWait = consumerInfo.config.ack_wait || defaultAckWait
+    const ackWait = consumerDefaults(def).ack_wait
     // Pull messages from the consumer
     const run = () => {
       ps.pull({ batch, expires: pullInterval })
@@ -115,11 +136,11 @@ export const jobProcessor = async (opts?: ConnectionOptions) => {
     for await (const msg of ps) {
       debug('received %O', msg.info)
       deferred = defer()
-      if (autoExtendAckTimeout) {
-        extendAckTimer = extendAckTimeout(ackWait, msg)
-      }
-
+      // Auto-extend ack timeout
+      const extendAckTimer =
+        autoExtendAckTimeout && extendAckTimeout(ackWait, msg)
       try {
+        // Process the message
         await def.perform(msg, { signal: abortController.signal, def, js })
         debug('completed')
         // Ack message
@@ -131,6 +152,7 @@ export const jobProcessor = async (opts?: ConnectionOptions) => {
         // Negative ack message with backoff
         msg.nak(backoffMs)
       } finally {
+        // Clear auto-extend timer
         if (extendAckTimer) {
           clearInterval(extendAckTimer)
         }
@@ -156,7 +178,9 @@ export const jobProcessor = async (opts?: ConnectionOptions) => {
    * ```
    */
   const stop = () => {
+    // Send abort signal to perform
     abortController.abort()
+    // Don't pull any more messages
     clearInterval(timer)
     return deferred?.promise
   }