+++ javascript/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.6
+# 0.1.0
 
 * Moved `jobScheduler` to [ha-job-scheduler](https://www.npmjs.com/package/ha-job-scheduler).
 * Add flag `jobDef.autoExtendAckTimeout` (set to true by default) to auto-extend message acknowledgement timeout